verified all the librairy so i removed the stddef.h from the header and change the type of my number_of_node in the 0-print_list.c

+++ singly_linked_lists/0-print_list.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-#include <stdlib.h>
 #include <stdio.h>
 
 /**
@@ -12,7 +11,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int number_of_node = 0;
+	size_t number_of_node = 0;
 
 	while (h != NULL)
 	{
